+++ CHANGELOG.md
@@ -5,7 +5,8 @@
 - optimizers now hold a `random_state` attribute which can be seeded (`optimizer.random_state.seed(12)`). Seeding `numpy`'s global random
   state seed **before** initializing the optimizer still works as well (but setting it between initialization and optimization does not while
   it used to work).
-- add a way to request a given `Candidate` from the `ask` method: `request` method (experimental feature).
+- added a `Scalar` variable as a shortcut to `Array(1).asscalar(dtype)` to simplify specifying instrumentation.
+- added a way to request a given `Candidate` from the `ask` method: `suggest` method (experimental feature, name and behavior may change).
 
 ## v0.2.1